Refactor reporting base classes into a common module
Issue: https://github.com/WorldconVotingSystems/nomnom/issues/199
Milestone: Seattle 2025 Nominations Close
